feat: implement Default trait for IncrementalCheckpoint and JsonGrammar
refactor: simplify QuantizationResolver logic and improve clarity

refactor: streamline InjectionDetector structure and methods

fix: optimize DefaultSampler's logit adjustments for clarity

fix: update callback closure in streaming tests for unused parameter

fix: improve WeightsVariant's architecture retrieval with cloned()

fix: enhance AtheerEngine memory management during pressure relief

fix: update FFI functions to ensure safety and null pointer handling

fix: add missing thermal state constants for iOS

refactor: improve GenericMonitor's default sampling method

refactor: enhance PowerState and ThermalState with Default trait

refactor: streamline HandoffPhase and HandoffProtocol implementations

fix: optimize KvCache entry insertion logic

fix: improve L1ActiveCache's entry removal logic

fix: enhance MemoryBank's VRAM pressure ratio calculation

refactor: clean up JsonGrammar's parsing logic

fix: optimize GrammarSampler's token validation loop

refactor: enhance InferenceMode with Default trait

fix: streamline Eco NGramCache's n-gram insertion logic

refactor: improve Orchestrator's mode change logic for clarity

refactor: clean up ThermalModel's code structure

chore: add Clippy fixes documentation for candle-core

chore: silence Clippy lints in candle-core library

chore: add Clippy fixes documentation for candle-coreml

chore: silence Clippy lints in candle-coreml library

chore: silence Clippy lints in candle-transformers library

fix: update landing page statistics to reflect accurate test counts

+++ atheer-memory-bank/src/memory_bank.rs
@@ -261,7 +261,7 @@ impl MemoryBank {
     /// VRAM pressure ratio (0.0–1.0+) relative to `max_size_mb`.
     /// Values > 1.0 indicate the cache exceeds its budget.
     pub fn vram_pressure_ratio(&self) -> f32 {
-        let max_bytes = (self.max_size_mb as f64 * 1_048_576.0) as f64;
+        let max_bytes = self.max_size_mb as f64 * 1_048_576.0;
         if max_bytes <= 0.0 {
             return 1.0;
         }
